LayoutTests/http/tests/site-isolation/pageswap-event.html

<!-- webkit-test-runner [ SiteIsolationEnabled=true ] -->
<!-- This test is adapted from LayoutTests/imported/w3c/web-platform-tests/html/browsers/browsing-the-web/history-traversal/pageswap/pageswap-cross-origin.sub.html -->

<!DOCTYPE html>
<title>Tests that pageswap fires when site isolation is enabled.</title>

<script src="/js-test-resources/js-test.js"></script>
<script>
description("Tests that pageswap fires when site isolation is enabled.");
jsTestIsAsync = true;

const params = new URLSearchParams(location.search);
// The page the popup navigates to.
const is_new_page = params.has('new');
// The initial page in the popup.
const is_popup_page = params.has('popup');
// The test page itself.
const is_test_page = !is_popup_page && !is_new_page;

if (is_test_page) {
  onload = () => {
    popup = window.open("?popup");
  };

  onmessage = (event) => {
    shouldBeEqualToString("event.data", "pageswap");
    popup.close();
    finishJSTest();
  };
}

if (is_popup_page) {
  onload = () => {
    requestAnimationFrame(() => requestAnimationFrame(() => {
      location.href = "http://localhost:8000/site-isolation/pageswap-event.html?new";
    }));
  };

  onpageswap = () => window.opener.postMessage("pageswap", "*");
}
</script>
