LayoutTests/http/tests/site-isolation/pageswap-event-expected.txt

Tests that pageswap fires when site isolation is enabled.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS event.data is "pageswap"
PASS successfullyParsed is true

TEST COMPLETE
